# specification/spec/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Help Desk API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: '"Help Desk" is a comprehensive customer support application designed to streamline and optimize customer service operations. With intuitive ticket management, efficient communication tools, and a robust knowledge base, Help Desk empowers organizations to provide prompt assistance and resolution to customer inquiries and issues. From ticket submission to resolution, Help Desk ensures a seamless support experience, enhancing customer satisfaction and loyalty.'
  contact:
    name: Pixfar Development Team
    url: https://pixfar.com
    email: anamul@pixfar.com
servers:
  - description: Local Development
    url: http://127.0.0.1:5000
  - description: Mock Server
    url: http://127.0.0.1:5055/_mock/openapi
...
paths:
...
  /api/v1/auth/forgot-password:
    post:
      description: route for reseting password if any user forgot password to login
      summary: forgot password
      operationId: forgot-password
      security: []
      tags:
        - Auth
      requestBody:
        $ref: '#/components/requestBodies/forgot-password'
      responses:
        '200':
          $ref: '#/components/responses/forgot-password'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /api/v1/auth/reset-password:
    post:
      description: this api url is used to reset the login credentials if any one forgot the login credentials
      summary: reset password
      operationId: reset-password
      security: []
      tags:
        - Auth
      requestBody:
        $ref: '#/components/requestBodies/reset-password'
      responses:
        '200':
          $ref: '#/components/responses/reset-password'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
  /api/v1/tickets/{id}:
    get:
      description: This api endpoint returns a ticket by id
      summary: Get a ticket by id
      operationId: get-ticket
      tags:
        - Ticket
      security:
        - JWT: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          $ref: '#/components/responses/ticket-by-id'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    patch:
      description: This api endpoint updates a ticket by id
      summary: Update a ticket by id
      operationId: update-ticket
      tags:
        - Ticket
      security:
        - JWT: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        $ref: '#/components/requestBodies/update-ticket'
      responses:
        '200':
          $ref: '#/components/responses/update-ticket'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    delete:
      description: This api endpoint deletes a ticket by id
      summary: Delete a ticket by id
      operationId: delete-ticket
      tags:
        - Ticket
      security:
        - JWT: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          $ref: '#/components/responses/delete-ticket'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
components:
  parameters:
...
    id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: cuid
        description: The unique identifier of the member.
        example: clvfmx5380001ngrve2rsnbv5
...
  schemas:
...
    forgot-password:
      type: object
      properties:
        email:
          type: string
          description: email which use to create the account
          example: example@gmail.com
      required:
        - email
    reset-password:
      type: object
      properties:
        token:
          type: string
          example: token
        email:
          type: string
          format: email
          example: example@gmail.com
        password:
          type: string
          example: password
        confirmPassword:
          type: string
          example: password
      required:
        - token
        - email
        - password
        - confirmPassword
...
    ticket:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: title
        content:
          type: string
          example: content
        system:
          type: string
          example: low
        category:
          type: string
          example: assignee
        submitter:
          type: string
          example: systemId
        assignedTeam:
          type: string
          example: reporter
        assignedUser:
          type: string
          example: '2021-01-01'
        status:
          type: string
          example: open
        created_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
      required:
        - title
        - content
        - system
        - category
        - submitter
        - status
...
    update-ticket:
      type: object
      properties:
        title:
          type: string
          example: title
        content:
          type: string
          example: content
        system:
          type: string
          example: low
        category:
          type: string
          example: assignee
        submitter:
          type: string
          example: systemId
        assignedTeam:
          type: string
          example: reporter
        assignedUser:
          type: string
          example: '2021-01-01'
        status:
          type: string
          example: open
...
  responses:
    '400':
      description: bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 400
              status:
                type: string
                example: error
              message:
                type: string
                example: Bad request
              errors:
                type: object
                properties:
                  field:
                    type: string
                    example: message
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 401
              status:
                type: string
                example: error
              message:
                type: string
                example: Unauthorized
    '403':
      description: forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 403
              status:
                type: string
                example: error
              message:
                type: string
                example: Forbidden
    '404':
      description: not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 404
              status:
                type: string
                example: error
              message:
                type: string
                example: Not found
    '500':
      description: internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 500
              status:
                type: string
                example: error
              message:
                type: string
                example: Internal server error
...
    forgot-password:
      description: reset email send successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 200
              status:
                type: string
                example: success
              message:
                type: string
                example: Reset password email send to successfully
    reset-password:
      description: success responses for reset password request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 200
              status:
                type: string
                example: success
              message:
                type: string
                example: password reset successful. Please login again.
...
    ticket-by-id:
      description: Request body for getting a ticket by ID
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 200
              status:
                type: string
                example: success
              message:
                type: string
                example: Ticket retrieved successfully
              data:
                $ref: '#/components/schemas/ticket'
    delete-ticket:
      description: Response body for deleting a ticket
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 200
              status:
                type: string
                example: success
              message:
                type: string
                example: Ticket deleted successfully
    update-ticket:
      description: Response body for updating a ticket
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 200
              status:
                type: string
                example: success
              message:
                type: string
                example: Ticket updated successfully
              data:
                type: object
                properties:
                  ticket:
                    $ref: '#/components/schemas/ticket'
...
  requestBodies:
...
    forgot-password:
      description: forgot password request body
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/forgot-password'
    reset-password:
      description: request body for reset the login credentials
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/reset-password'
...
    update-ticket:
      description: Request body for updating a ticket
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/update-ticket'
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT